macos: make SurfaceView cross-platform!
